Change potion effect time from ticks to kotlin duration
