+++ src/main/kotlin/io/github/dockyardmc/implementations/commands/DefaultCommands.kt
@@ -16,6 +16,7 @@ class DefaultCommands: DefaultImplementationModule {
         SchedulerCommand()
         ClearCommand()
         ListCommand()
+        EffectCommand()
         if(ConfigManager.config.debug) {
             DebugCommand()
         }
