+++ src/main/kotlin/io/github/dockyardmc/scheduler/runnables/AsyncTimedSequence.kt
@@ -28,3 +28,8 @@ class AsyncTimedSequence() {
 val Int.ticks get() = (this * 50).toDuration(DurationUnit.MILLISECONDS)
 
 fun Int.ticks(): Duration = (this * 50).toDuration(DurationUnit.MILLISECONDS)
+
+val Duration.inWholeMinecraftTicks: Int get() {
+    if(this.isInfinite()) return -1
+    return (this.inWholeMilliseconds / 50).toInt()
+}
